Added default current location button, removed our fab for that, made search overlay work, deleted LocSearchActivity, and modified notification action
