+++ app/src/main/AndroidManifest.xml
@@ -42,13 +42,13 @@
             android:name=".MapsActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LocSearchActivity" />
         <activity
             android:name=".LocSettingsActivity"
             android:label="@string/title_activity_add_loc_settings"
