+++ app/src/main/java/com/visual/android/locsilence/CustomProximityMap.java
@@ -3,16 +3,14 @@ package com.visual.android.locsilence;
 import android.Manifest;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.location.Criteria;
 import android.location.LocationManager;
+import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
@@ -23,9 +21,6 @@ import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.MarkerOptions;
-
-import java.util.List;
 
 public class CustomProximityMap extends AppCompatActivity implements OnMapReadyCallback {
 
@@ -52,6 +47,7 @@ public class CustomProximityMap extends AppCompatActivity implements OnMapReadyC
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
                 .findFragmentById(R.id.map_customProxy);
         mapFragment.getMapAsync(this);
+
 
         // Init listeners
         find_selectedLoc.setOnClickListener(new View.OnClickListener() {
@@ -127,6 +123,7 @@ public class CustomProximityMap extends AppCompatActivity implements OnMapReadyC
                 if (mMap != null) {
                     mMap.setMyLocationEnabled(true);
                     mMap.getUiSettings().setMyLocationButtonEnabled(true);
+                    mMap.getUiSettings().setCompassEnabled(true);
                 }
             }
         }
