+++ app/src/main/java/com/visual/android/locsilence/LocSettingsActivity.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class LocSettingsActivity extends AppCompatActivity {
 
     int PLACE_AUTOCOMPLETE_REQUEST_CODE = 1;
-    private static final String TAG = LocSearchActivity.class.getSimpleName();
+    private static final String TAG = LocSettingsActivity.class.getSimpleName();
     String[] volumeTypes = {"Ringtone", "Notifications", "Alarms"};
     final int DEFAULT_RADIUS = 100;
